Add Perl script to convert ipmp_ping output to CSV
This script converts the output of ipmp_ping to a CSV
file, which can be imported into statistics software.

One line in the CSV output is one reply packet.

See the file header for the options I use to generate my
IPMP output. The output uses semicola (;) for seperating
fields and dots (.) for the decimal places of numbers.

The complete paths, the timestamps of the paths and the
flow counters are serialized to fit into single fields,
respectively, with lines seperated by a '>'.

The first line of the output is a header for the CSV file:
src_ip;dst_ip;fwd_path_hops;rev_path_hops;rtt;full_path;timestamps;flow_counter

Have fun.

Signed-off-by: Florian Sesser <sesser@in.tum.de>